+++ src/ct_physics.jl
@@ -751,7 +751,7 @@ function StimulatedRecombination(u, node, data)
 end
 
 
-function addRecombination!(f, u, node, data, ::SRHWithoutTrapsType)
+function addRecombination!(f, u, node, data, ::Val{true})
 
     params = data.params
     ireg = node.region
@@ -792,6 +792,9 @@ function addRecombination!(f, u, node, data, ::SRHWithoutTrapsType)
     return nothing
 end
 
+function addRecombination!(f, u, node, data, ::Val{false})
+    return nothing
+end
 
 function addStimulatedRecombination!(f, u, node, data, ::Type{LaserModelOff})
     return nothing
@@ -877,7 +880,7 @@ Function which builds right-hand side of electric charge carriers.
 function RHSContinuityEquations!(f, u, node, data)
 
     # dependent on user information concerncing recombination
-    addRecombination!(f, u, node, data, data.bulkRecombination.bulk_recomb_SRH)
+    addRecombination!(f, u, node, data, Val(data.bulkRecombination.bulk_recomb))
     # dependent on user information concerning laser model
     addStimulatedRecombination!(f, u, node, data, data.laserModel)
     # dependent on user information concerncing generation